chore(vlms): move vision implementations to standalone llm packages
