+++ pkgs/community/swarmauri_llm_groq/swarmauri_llm_groq/__init__.py
@@ -2,8 +2,15 @@ from .GroqAIAudio import GroqAIAudio
 from .GroqModel import GroqModel
 from .GroqToolModel import GroqToolModel
 from .GroqVisionModel import GroqVisionModel
+from .GroqVLM import GroqVLM
 
-__all__ = ["GroqAIAudio", "GroqModel", "GroqToolModel", "GroqVisionModel"]
+__all__ = [
+    "GroqAIAudio",
+    "GroqModel",
+    "GroqToolModel",
+    "GroqVisionModel",
+    "GroqVLM",
+]
 
 try:
     from importlib.metadata import PackageNotFoundError, version
