+++ pkgs/swarmauri_standard/swarmauri_standard/vlms/GroqVLM.py
@@ -1,437 +1,19 @@
-import asyncio
-import json
-from typing import (
-    Any,
-    AsyncGenerator,
-    Dict,
-    Generator,
-    List,
-    Literal,
-    Optional,
-    Type,
+"""Compatibility shims for deprecated swarmauri_standard.vlms locations.
+
+VLM implementations now live in their dedicated provider packages.
+"""
+
+import warnings
+
+from swarmauri_llm_groq import GroqVLM as _GroqVLM
+
+warnings.warn(
+    "Importing GroqVLM from swarmauri_standard.vlms is deprecated. "
+    "Use swarmauri_llm_groq.GroqVLM instead.",
+    DeprecationWarning,
+    stacklevel=2,
 )
 
-import httpx
-from pydantic import PrivateAttr, SecretStr
-from swarmauri_base.ComponentBase import ComponentBase
-from swarmauri_base.messages.MessageBase import MessageBase
-from swarmauri_base.vlms.VLMBase import VLMBase
+GroqVLM = _GroqVLM
 
-from swarmauri_standard.conversations.Conversation import Conversation
-from swarmauri_standard.messages.AgentMessage import AgentMessage, UsageData
-from swarmauri_standard.utils.retry_decorator import retry_on_status_codes
-
-
-@ComponentBase.register_type(VLMBase, "GroqVLM")
-class GroqVLM(VLMBase):
-    """
-    GroqVCM class for interacting with the Groq vision language models API.
-    This class
-    provides synchronous and asynchronous methods to send conversation data to
-    the
-    model, receive predictions, and stream responses.
-
-    Attributes:
-        api_key (str): API key for authenticating requests to the Groq API.
-        allowed_models (List[str]): List of allowed model names that can be
-        used.
-        name (str): The default model name to use for predictions.
-        type (Literal["GroqModel"]): The type identifier for this class.
-
-    Allowed Models resources: https://console.groq.com/docs/models
-    """
-
-    api_key: SecretStr
-    allowed_models: List[str] = [
-        "meta-llama/llama-4-scout-17b-16e-instruct",
-        "meta-llama/llama-4-maverick-17b-128e-instruct",
-    ]
-    name: str = "meta-llama/llama-4-scout-17b-16e-instruct"
-    type: Literal["GroqVLM"] = "GroqVLM"
-    _client: httpx.Client = PrivateAttr(default=None)
-    _async_client: httpx.AsyncClient = PrivateAttr(default=None)
-    _BASE_URL: str = PrivateAttr(
-        default="https://api.groq.com/openai/v1/chat/completions"
-    )
-
-    def __init__(self, **data: dict[str, Any]):
-        """
-        Initialize the GroqAIAudio class with the provided data.
-
-        Args:
-            **data: Arbitrary keyword arguments containing initialization data.
-        """
-        super().__init__(**data)
-        self._client = httpx.Client(
-            headers={
-                "Authorization": f"Bearer {self.api_key.get_secret_value()}"
-            },
-            base_url=self._BASE_URL,
-        )
-        self._async_client = httpx.AsyncClient(
-            headers={
-                "Authorization": f"Bearer {self.api_key.get_secret_value()}"
-            },
-            base_url=self._BASE_URL,
-        )
-
-    def _format_messages(
-        self,
-        messages: List[Type[MessageBase]],
-    ) -> List[Dict[str, Any]]:
-        """
-        Formats conversation messages into the structure expected by the API.
-
-        Args:
-            messages (List[MessageBase]): List of message objects from the
-            conversation history.
-
-        Returns:
-            List[Dict[str, Any]]: List of formatted message dictionaries.
-        """
-
-        formatted_messages = []
-        for message in messages:
-            formatted_message = message.model_dump(
-                include=["content", "role", "name"], exclude_none=True
-            )
-
-            if isinstance(formatted_message["content"], list):
-                formatted_message["content"] = [
-                    {"type": item["type"], **item}
-                    for item in formatted_message["content"]
-                ]
-
-            formatted_messages.append(formatted_message)
-        return formatted_messages
-
-    def _prepare_usage_data(self, usage_data) -> UsageData:
-        """
-        Prepares and validates usage data received from the API response.
-
-        Args:
-            usage_data (dict): Raw usage data from the API response.
-
-        Returns:
-            UsageData: Validated usage data instance.
-        """
-        return UsageData.model_validate(usage_data)
-
-    @retry_on_status_codes((429, 529), max_retries=1)
-    def predict_vision(
-        self,
-        conversation: Conversation,
-        temperature: float = 0.7,
-        max_tokens: int = 256,
-        top_p: float = 1.0,
-        enable_json: bool = False,
-        stop: Optional[List[str]] = None,
-    ) -> Conversation:
-        """
-        Generates a response from the model based on the given conversation.
-
-        Args:
-            conversation (Conversation): Conversation object with message
-            history.
-            temperature (float): Sampling temperature for response diversity.
-            max_tokens (int): Maximum tokens for the model's response.
-            top_p (float): Cumulative probability for nucleus sampling.
-            enable_json (bool): Whether to format the response as JSON.
-            stop (Optional[List[str]]): List of stop sequences for response
-            termination.
-
-        Returns:
-            Conversation: Updated conversation with the model's response.
-        """
-        formatted_messages = self._format_messages(conversation.history)
-        payload = {
-            "model": self.name,
-            "messages": formatted_messages,
-            "temperature": temperature,
-            "max_tokens": max_tokens,
-            "top_p": top_p,
-            "stop": stop or [],
-        }
-        if enable_json:
-            payload["response_format"] = "json_object"
-
-        response = self._client.post(self._BASE_URL, json=payload)
-
-        response.raise_for_status()
-
-        response_data = response.json()
-
-        message_content = response_data["choices"][0]["message"]["content"]
-        usage_data = response_data.get("usage", {})
-
-        usage = self._prepare_usage_data(usage_data)
-        conversation.add_message(
-            AgentMessage(content=message_content, usage=usage)
-        )
-        return conversation
-
-    @retry_on_status_codes((429, 529), max_retries=1)
-    async def apredict_vision(
-        self,
-        conversation: Conversation,
-        temperature: float = 0.7,
-        max_tokens: int = 256,
-        top_p: float = 1.0,
-        enable_json: bool = False,
-        stop: Optional[List[str]] = None,
-    ) -> Conversation:
-        """
-        Async method to generate a response from the model based on the given
-        conversation.
-
-        Args:
-            conversation (Conversation): Conversation object with message
-            history.
-            temperature (float): Sampling temperature for response diversity.
-            max_tokens (int): Maximum tokens for the model's response.
-            top_p (float): Cumulative probability for nucleus sampling.
-            enable_json (bool): Whether to format the response as JSON.
-            stop (Optional[List[str]]): List of stop sequences for response
-            termination.
-
-        Returns:
-            Conversation: Updated conversation with the model's response.
-        """
-        formatted_messages = self._format_messages(conversation.history)
-        payload = {
-            "model": self.name,
-            "messages": formatted_messages,
-            "temperature": temperature,
-            "max_tokens": max_tokens,
-            "top_p": top_p,
-            "stop": stop or [],
-        }
-        if enable_json:
-            payload["response_format"] = "json_object"
-
-        response = await self._async_client.post(self._BASE_URL, json=payload)
-        response.raise_for_status()
-
-        response_data = response.json()
-
-        message_content = response_data["choices"][0]["message"]["content"]
-        usage_data = response_data.get("usage", {})
-
-        usage = self._prepare_usage_data(usage_data)
-        conversation.add_message(
-            AgentMessage(content=message_content, usage=usage)
-        )
-        return conversation
-
-    @retry_on_status_codes((429, 529), max_retries=1)
-    def stream(
-        self,
-        conversation: Conversation,
-        temperature: float = 0.7,
-        max_tokens: int = 256,
-        top_p: float = 1.0,
-        enable_json: bool = False,
-        stop: Optional[List[str]] = None,
-    ) -> Generator[str, None, None]:
-        """
-        Streams response text from the model in real-time.
-
-        Args:
-            conversation (Conversation): Conversation object with message
-            history.
-            temperature (float): Sampling temperature for response diversity.
-            max_tokens (int): Maximum tokens for the model's response.
-            top_p (float): Cumulative probability for nucleus sampling.
-            enable_json (bool): Whether to format the response as JSON.
-            stop (Optional[List[str]]): List of stop sequences for response
-            termination.
-
-        Yields:
-            str: Partial response content from the model.
-        """
-
-        formatted_messages = self._format_messages(conversation.history)
-        payload = {
-            "model": self.name,
-            "messages": formatted_messages,
-            "temperature": temperature,
-            "max_tokens": max_tokens,
-            "top_p": top_p,
-            "stream": True,
-            "stop": stop or [],
-        }
-        if enable_json:
-            payload["response_format"] = "json_object"
-
-        response = self._client.post(self._BASE_URL, json=payload)
-
-        response.raise_for_status()
-        message_content = ""
-        for line in response.iter_lines():
-            json_str = line.replace("data: ", "")
-            try:
-                if json_str:
-                    chunk = json.loads(json_str)
-                    if chunk["choices"][0]["delta"]:
-                        delta = chunk["choices"][0]["delta"]["content"]
-                        message_content += delta
-                        yield delta
-            except json.JSONDecodeError:
-                pass
-
-        conversation.add_message(AgentMessage(content=message_content))
-
-    @retry_on_status_codes((429, 529), max_retries=1)
-    async def astream(
-        self,
-        conversation: Conversation,
-        temperature: float = 0.7,
-        max_tokens: int = 256,
-        top_p: float = 1.0,
-        enable_json: bool = False,
-        stop: Optional[List[str]] = None,
-    ) -> AsyncGenerator[str, None]:
-        """
-        Async generator that streams response text from the model in real-time.
-
-        Args:
-            conversation (Conversation): Conversation object with message
-            history.
-            temperature (float): Sampling temperature for response diversity.
-            max_tokens (int): Maximum tokens for the model's response.
-            top_p (float): Cumulative probability for nucleus sampling.
-            enable_json (bool): Whether to format the response as JSON.
-            stop (Optional[List[str]]): List of stop sequences for response
-            termination.
-
-        Yields:
-            str: Partial response content from the model.
-        """
-
-        formatted_messages = self._format_messages(conversation.history)
-        payload = {
-            "model": self.name,
-            "messages": formatted_messages,
-            "temperature": temperature,
-            "max_tokens": max_tokens,
-            "top_p": top_p,
-            "stream": True,
-            "stop": stop or [],
-        }
-        if enable_json:
-            payload["response_format"] = "json_object"
-
-        response = await self._async_client.post(self._BASE_URL, json=payload)
-
-        response.raise_for_status()
-        message_content = ""
-        async for line in response.aiter_lines():
-            json_str = line.replace("data: ", "")
-            try:
-                if json_str:
-                    chunk = json.loads(json_str)
-                    if chunk["choices"][0]["delta"]:
-                        delta = chunk["choices"][0]["delta"]["content"]
-                        message_content += delta
-                        yield delta
-            except json.JSONDecodeError:
-                pass
-
-        conversation.add_message(AgentMessage(content=message_content))
-
-    def batch(
-        self,
-        conversations: List[Conversation],
-        temperature: float = 0.7,
-        max_tokens: int = 256,
-        top_p: float = 1.0,
-        enable_json: bool = False,
-        stop: Optional[List[str]] = None,
-    ) -> List[Conversation]:
-        """
-        Processes a batch of conversations and generates responses for each
-        sequentially.
-
-        Args:
-            conversations (List[Conversation]): List of conversations to
-            process.
-            temperature (float): Sampling temperature for response diversity.
-            max_tokens (int): Maximum tokens for each response.
-            top_p (float): Cumulative probability for nucleus sampling.
-            enable_json (bool): Whether to format the response as JSON.
-            stop (Optional[List[str]]): List of stop sequences for response
-            termination.
-
-        Returns:
-            List[Conversation]: List of updated conversations with model
-            responses.
-        """
-        results = []
-        for conversation in conversations:
-            result_conversation = self.predict_vision(
-                conversation,
-                temperature=temperature,
-                max_tokens=max_tokens,
-                top_p=top_p,
-                enable_json=enable_json,
-                stop=stop,
-            )
-            results.append(result_conversation)
-        return results
-
-    async def abatch(
-        self,
-        conversations: List[Conversation],
-        temperature: float = 0.7,
-        max_tokens: int = 256,
-        top_p: float = 1.0,
-        enable_json: bool = False,
-        stop: Optional[List[str]] = None,
-        max_concurrent=5,
-    ) -> List[Conversation]:
-        """
-        Async method for processing a batch of conversations concurrently.
-
-        Args:
-            conversations (List[Conversation]): List of conversations to
-            process.
-            temperature (float): Sampling temperature for response diversity.
-            max_tokens (int): Maximum tokens for each response.
-            top_p (float): Cumulative probability for nucleus sampling.
-            enable_json (bool): Whether to format the response as JSON.
-            stop (Optional[List[str]]): List of stop sequences for response
-            termination.
-            max_concurrent (int): Maximum number of concurrent requests.
-
-        Returns:
-            List[Conversation]: List of updated conversations with model
-            responses.
-        """
-        semaphore = asyncio.Semaphore(max_concurrent)
-
-        async def process_conversation(conv: Conversation) -> Conversation:
-            async with semaphore:
-                return await self.apredict_vision(
-                    conv,
-                    temperature=temperature,
-                    max_tokens=max_tokens,
-                    top_p=top_p,
-                    enable_json=enable_json,
-                    stop=stop,
-                )
-
-        tasks = [process_conversation(conv) for conv in conversations]
-        return await asyncio.gather(*tasks)
-
-    def get_allowed_models(self) -> List[str]:
-        """
-        Queries the LLMProvider API endpoint to retrieve the list of allowed
-        models.
-
-        Returns:
-            List[str]: List of allowed model names.
-        """
-        return [
-            "meta-llama/llama-4-scout-17b-16e-instruct",
-            "meta-llama/llama-4-maverick-17b-128e-instruct",
-        ]
+__all__ = ["GroqVLM"]
